+++ frontend/src/components/SimulacionOperacionPrivada.tsx
@@ -6,7 +6,7 @@ import {
 import { type SimulacionResponse } from "../types/Simulacion";
 import { getClienteActual } from "../services/usuarioService";
 import { type Cliente } from "../types/Cliente";
-import { type MetodoFinanciero } from "../types/MetodoFinanciero";
+import { type MetodoFinanciero } from "../features/financiero/types/MetodoFinanciero";
 import { getDivisasConTasa } from "../services/divisaService";
 import { type Divisa } from "../types/Divisa";
 import type { DecodedToken } from "../types/User";
