+++ frontend/src/features/financiero/components/MetodoFinancieroCard.tsx
@@ -13,7 +13,7 @@ import type {
   BilleteraDigital,
   Tarjeta,
 } from "../types/MetodoFinanciero";
-import { BILLETERAS_DIGITALES, BILLETERAS_DIGITALES_CATALOGO, CUENTAS_BANCARIAS } from "../../../types/perms";
+import { BILLETERAS_DIGITALES, CUENTAS_BANCARIAS } from "../../../types/perms";
 import Can from "../../../components/Can";
 
 type MetodoFinancieroItem = CuentaBancaria | BilleteraDigital | Tarjeta;
@@ -29,7 +29,7 @@ interface MetodoFinancieroCardProps {
   onToggleActive: (item: MetodoFinancieroItem) => void;
 }
 
-export const MetodoFinancieroCard: React.FC<MetodoFinancieroCardProps> = ({
+const MetodoFinancieroCard: React.FC<MetodoFinancieroCardProps> = ({
   item,
   onView,
   onEdit,
@@ -48,28 +48,6 @@ export const MetodoFinancieroCard: React.FC<MetodoFinancieroCardProps> = ({
     }
   };
 
-  const getChangePerm = () => {
-    switch (item.tipo) {
-      case "cuentas":
-        return [CUENTAS_BANCARIAS.CHANGE];
-      case "billeteras digitales":
-        return [BILLETERAS_DIGITALES.CHANGE];
-      default:
-        return [];
-    }
-  };
-
-  const getDeletePerm = () => {
-    switch (item.tipo) {
-      case "cuentas":
-        return [CUENTAS_BANCARIAS.DELETE];
-      case "billeteras digitales":
-        return [BILLETERAS_DIGITALES.DELETE];
-      default:
-        return [];
-    }
-  };
-
   const getTitle = () => {
     switch (item.tipo) {
       case "cuentas":
@@ -77,10 +55,7 @@ export const MetodoFinancieroCard: React.FC<MetodoFinancieroCardProps> = ({
         return cuenta.banco_nombre || `Banco ID: ${cuenta.banco}`;
       case "billeteras digitales":
         const billetera = item as BilleteraDigital;
-        return (
-          billetera.plataforma_nombre ||
-          `Plataforma ID: ${billetera.plataforma}`
-        );
+        return billetera.plataforma_nombre || `Plataforma ID: ${billetera.plataforma}`;
       case "tarjetas":
         return `**** ${(item as Tarjeta).last4}`;
       default:
@@ -95,18 +70,12 @@ export const MetodoFinancieroCard: React.FC<MetodoFinancieroCardProps> = ({
         return (
           <div className="space-y-2">
             <div>
-              <span className="text-sm font-medium text-gray-700">
-                Titular:{" "}
-              </span>
+              <span className="text-sm font-medium text-gray-700">Titular: </span>
               <span className="text-sm text-gray-900">{cuenta.titular}</span>
             </div>
             <div>
-              <span className="text-sm font-medium text-gray-700">
-                CBU/CVU:{" "}
-              </span>
-              <span className="text-sm text-gray-900">
-                {cuenta.cbu_cvu.slice(-4)}
-              </span>
+              <span className="text-sm font-medium text-gray-700">CBU/CVU: </span>
+              <span className="text-sm text-gray-900">***{cuenta.cbu_cvu.slice(-4)}</span>
             </div>
           </div>
         );
@@ -115,21 +84,13 @@ export const MetodoFinancieroCard: React.FC<MetodoFinancieroCardProps> = ({
         return (
           <div className="space-y-2">
             <div>
-              <span className="text-sm font-medium text-gray-700">
-                Usuario:{" "}
-              </span>
-              <span className="text-sm text-gray-900">
-                {billetera.usuario_id || billetera.email}
-              </span>
+              <span className="text-sm font-medium text-gray-700">Usuario: </span>
+              <span className="text-sm text-gray-900">{billetera.usuario_id || billetera.email}</span>
             </div>
             {billetera.telefono && (
               <div>
-                <span className="text-sm font-medium text-gray-700">
-                  Teléfono:{" "}
-                </span>
-                <span className="text-sm text-gray-900">
-                  {billetera.telefono}
-                </span>
+                <span className="text-sm font-medium text-gray-700">Teléfono: </span>
+                <span className="text-sm text-gray-900">{billetera.telefono}</span>
               </div>
             )}
           </div>
@@ -139,18 +100,13 @@ export const MetodoFinancieroCard: React.FC<MetodoFinancieroCardProps> = ({
         return (
           <div className="space-y-2">
             <div>
-              <span className="text-sm font-medium text-gray-700">
-                Titular:{" "}
-              </span>
+              <span className="text-sm font-medium text-gray-700">Titular: </span>
               <span className="text-sm text-gray-900">{tarjeta.titular}</span>
             </div>
             <div>
-              <span className="text-sm font-medium text-gray-700">
-                Vencimiento:{" "}
-              </span>
+              <span className="text-sm font-medium text-gray-700">Vencimiento: </span>
               <span className="text-sm text-gray-900">
-                {tarjeta.exp_month.toString().padStart(2, "0")}/
-                {tarjeta.exp_year}
+                {tarjeta.exp_month.toString().padStart(2, '0')}/{tarjeta.exp_year}
               </span>
             </div>
           </div>
@@ -160,36 +116,62 @@ export const MetodoFinancieroCard: React.FC<MetodoFinancieroCardProps> = ({
     }
   };
 
+  const getChangePerm = () => {
+    switch (item.tipo) {
+      case "cuentas":
+        return [CUENTAS_BANCARIAS.change];
+      case "billeteras digitales":
+        return [BILLETERAS_DIGITALES.change];
+      default:
+        return [];
+    }
+  };
+
+  const getDeletePerm = () => {
+    switch (item.tipo) {
+      case "cuentas":
+        return [CUENTAS_BANCARIAS.delete];
+      case "billeteras digitales":
+        return [BILLETERAS_DIGITALES.delete];
+      default:
+        return [];
+    }
+  };
+
   return (
-    <div
-      className={`bg-white rounded-lg shadow-md border transition-all hover:shadow-lg ${
-        !item.is_active ? "opacity-60 border-gray-300" : "border-gray-200"
-      }`}
-    >
+    <div className="bg-white rounded-lg shadow-md border border-gray-200 hover:shadow-lg transition-all duration-200">
       <div className="p-4">
-        {/* Header: Ícono + Nombre principal + Badge de estado */}
-        <div className="flex items-start justify-between mb-4">
+        {/* Header: Status + Icon + Title */}
+        <div className="flex items-center justify-between mb-3">
           <div className="flex items-center space-x-3">
             {getIcon()}
-            <div className="flex-1 min-w-0">
-              <h3 className="text-lg font-semibold text-gray-900 truncate">
-                {getTitle()}
-              </h3>
+            <div>
+              <h3 className="text-lg font-semibold text-gray-900">{getTitle()}</h3>
             </div>
           </div>
-          <span
-            className={`inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium ${
-              item.is_active
-                ? "bg-green-100 text-green-800"
-                : "bg-red-100 text-red-800"
-            }`}
-          >
-            {item.is_active ? "Activo" : "Inactivo"}
-          </span>
+          <div className="flex items-center">
+            <span
+              className={`inline-flex px-2 py-1 text-xs font-semibold rounded-full ${
+                item.is_active
+                  ? "bg-green-100 text-green-800"
+                  : item.desactivado_por_catalogo
+                  ? "bg-red-100 text-red-800"
+                  : "bg-yellow-100 text-yellow-800"
+              }`}
+            >
+              {item.is_active
+                ? "Activo"
+                : item.desactivado_por_catalogo
+                ? "Desactivado por catálogo"
+                : "Inactivo"}
+            </span>
+          </div>
         </div>
 
         {/* Cuerpo: Información secundaria relevante */}
-        <div className="mb-4">{getBodyInfo()}</div>
+        <div className="mb-4">
+          {getBodyInfo()}
+        </div>
 
         {/* Footer: Acciones */}
         <div className="flex items-center justify-start space-x-3 pt-3 border-t border-gray-100">
@@ -213,7 +195,7 @@ export const MetodoFinancieroCard: React.FC<MetodoFinancieroCardProps> = ({
             </button>
           </Can>
 
-          <Can anyOf={item.is_active? getDeletePerm() : getChangePerm()}>
+          <Can anyOf={item.is_active ? getDeletePerm() : getChangePerm()}>
             <button
               onClick={() => onToggleActive(item)}
               disabled={!item.is_active && item.desactivado_por_catalogo}
