+++ frontend/src/routes/Routes.tsx
@@ -7,8 +7,8 @@ import RegisterPage from "../pages/RegisterPage";
 import RolesPage from "../pages/RolesPage";
 import MainMenuPage from "../pages/MenuPage";
 import DivisasPage from "../pages/DivisasPage";
-import MetodosFinancierosPage from "../pages/MetodosFinancierosPage";
-import MetodosFinancierosClientePage from "../pages/MetodosFinancierosClientePage";
+import MetodosFinancierosPage from "../features/financiero/Pages/MetodosFinancierosPage";
+import MetodosFinancierosClientePage from "../features/financiero/Pages/MetodosFinancierosClientePage";
 import ConfiguracionesPage from "../pages/ConfiguracionesPage";
 import CategoriaClientePage from "../pages/CategoriaClientePage";
 import {
