+++ frontend/src/types/Simulacion.ts
@@ -2,7 +2,7 @@
  * @fileoverview Tipos de datos para simulaciones de operaciones.
  */
 
-import {type MetodoFinanciero} from './MetodoFinanciero'
+import {type MetodoFinanciero} from '../features/financiero/types/MetodoFinanciero'
 
 /**
  * @typedef {Object} SimulacionRequest
